VCE_CESM/requirements.txt
pydicom>=2.2.0
openpyxl==3.0.10
numpy>=1.19.2
nibabel==4.0.2
matplotlib==3.6.2
Pillow==9.3.0
pandas>=1.2.0
torch>=1.9.0
opencv-python==4.6.0.66
torchvision>=0.10.0
PyYAML==6.0
scikit-learn==1.1.3
config==0.5.1
tqdm==4.64.1
albumentations>=1.0.0
imutils==0.5.4
scipy==1.9.3
seaborn==0.12.1
tensorflow==2.10.0
SimpleITK==2.2.0
toolz==0.12.0
scikit-image==0.19.3